requirements.txt
pep8
Django==1.11.5
django-grappelli==2.10.1
django-devutils
gunicorn
psycopg2
dj-database-url
django-import-export
sphinx
sphinx_rtd_theme
django-extensions
